requirements.txt
fastapi
uvicorn[standard]
web3
python-dotenv
firebase-admin
google-cloud-secret-manager